+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 
 ### Fixes
+ 
+- Correctly track and send GPU frame render data in profiles (#2823)
+
+## 8.3.3
+
+### Fixes
 
 - View hierarchy not sent for crashes (#2781)
 - Crash in Tracer for idle timeout (#2834)